+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## master
+* fix: `kumactl apply -v ...` support dots in variables name
+  [#636](https://github.com/Kong/kuma/pull/636)
 * feat: read only cached manager
   [#634](https://github.com/Kong/kuma/pull/634)
 * fix: explicitly set parameters in securityContext of kuma-init